Dockerfile: Install script-security plugin explicitly.
Without this, as of July 2020, a default installation of Jenkins
inside the Docker container will have problems.

The junit plugin depends on the script-security plugin, and will
automatically install a plugin version that is incompatible with
other plugins.  The result is that 'ftc add-nodes -b <board>'
results in an exception in ftc, from an error returned by Jenkins.

Fix the issue by installing the script-security plugin v1.68 before
all other plugins.

Signed-off-by: Laszlo Sitzer <laszlo.sitzer@native-instruments.de>