terminology: Use correct terminology for docker bind mounts
Fuego used "volume mounts" in a few places to describe the
mounts that it uses from directories in the host filesystem
to the Fuego container filesystem.  The correct terminology,
according to docker documentation is "bind mounts".

"volume mounts" is used by docker documentation to describe
something related, but different (a mount that is managed
by docker and not stricly tied to the host filesystem).

Use the correct docker terminology to avoid confusion.

Signed-off-by: Tim Bird <tim.bird@sony.com>