Dockerfile: resolve yet-another Jenkins plugin dependency issue
Fix a problem with dependencies on the junit and struct Jenkins
plugins.  Move struct to version 1.20 (from version 1.19), and move
it sooner in the plugin list.  Also, move junit 1.27 sooner in the
plugin list.

The result of getting the wrong plugins is usually that some other
plugins won't load.  Often, the first visibl sign of this is that
the description-setter plugin doesn't load, and the links to
the test log and run.json file are missing in the Jenkins user
interface.

The list of plugins fed to install-plugins.sh was alphabetical.
Don't do that anymore.  Make sure to load plugins that are dependend
on by others first, so that we get the versions we want (and not
something automatically pulled in by install-plugins.sh).

Unfortunately, the Jenkins version and plugin list is aging, fragile
and insecure.  It may be time for an upgrade.

Signed-off-by: Tim Bird <tim.bird@sony.com>