docs: Adjust label for Adding a Board
Pooja was using 'addingboard' as the label for this section.
I originally adjusted it to 'adding_a_board' but that caused
all kinds of trouble (with links not working on html pages).
Presumably this was because it somehow conflicted with the
section name itself?  (that's a guess).

Anyway, revert back to something with underscores: 'adding_board'

Also sneak in some other fixes:  Uppercase the word 'board'
in the filename.  Fix a heading error.  Word wrap some stuff.
The usual.

Signed-off-by: Tim Bird <tim.bird@sony.com>