docs: Architecture.rst - make table wider
It seemed to squishy.  Actually, I was trying to see if a wider table
would allow me to insert some spaces to allow better wrapping of table
data on the HTML version of the table, so that the table did not
require a horizontal scrollbar.  It didn't seem to have
any effect on the output.

Signed-off-by: Tim Bird <tim.bird@sony.com>