docs: Add some Sandbox pages, to test rst links and formatting
These pages are temporary, to determine what features are supported
by rst, and how best to express them.  There are lots of ways
to put a link in the document that are wrong and don't work.

Also, I plan to use these pages for experimenting with customizing
the sphinx_rtd_theme.

They should be hidden in the actual docs, if possible (or removed
when the rst work is completed).

Signed-off-by: Tim Bird <tim.bird@sony.com>