docs: remove trailing white space on many files
Many rst pages have trailing whitespace at the end of
lines in the files.  I removed them with a perl command.
This kind of commit is a pain for later 'git annotate'
commands, because this kind of change touches lots
of lines and masks viewing substantive changes to the
lines in the files.  However, it's nice to get to a
clean slate on the whitespace issue.

Signed-off-by: Tim Bird <tim.bird@sony.com>