docs: Core_interfaces: add section on jenkins python module
Add a section on Fuego's use of the jenkins python module.
Make a mental note about ftc using jenkins-cli.jar for create-job,
but the jenkins module for all other jenkins operations.  That's
weird.

While here fix up a couple of issues with the translation from
the wiki page.  Also, adjust a couple of incorrect definitions.

Signed-off-by: Tim Bird <tim.bird@sony.com>