Docker: upgrade Jenkins to 2.249.3 version
Adjust Dockerfile to use a newer version of Jenkins.  Apparently, the
Jenkins package server no longer provides the 2.164.2 version of the
jenkins debian package.  Change the mechanism for updating the 'flot'
plugin to avoid the use of jenkins-cli.jar.  Also, make the plugin
list generic (don't specify a fixed version for each plugin.)

Signed-off-by: Tim Bird <tim.bird@sony.com>