Change version to 1.5.8
This is an important point release of Fuego, that changes
the Jenkins version from 2.164.2 to 2.235.3.  The 2.164.2
Jenkins package is no longer available, so it it not an option
to use it anymore.  Fuego is fundamentally broken until this
is fixed and pushed to master.  This warrants a rather large
jump in the release revision for Fuego.

Version 1.6 of Fuego is still in progress, while some issues
are being hashed out, and I didn't want to wait for that release
to complete before getting this needed Jenkins fix out.

Signed-off-by: Tim Bird <tim.bird@sony.com>