Remove tweak to ttc to force use of /fuego-ro/ttc.conf
Version 2.0.2 of ttc now looks for /fuego-ro/ttc.conf directly,
and does not need to be modified at install time with this change.
Arguably, we should have been using TTC_CONF to solve this
problem anyway.  I'm leaving the line in the Dockerfile as
a comment for future reference (for a while).

Signed-off-by: Tim Bird <tim.bird@sony.com>