Add support for /usr/test directory in containter creation
If the host has a /usr/test directory, allow the container to see this
by mounting it as a volume mount inside the Fuego container.  This
allows some scripts on the host to be visible inside the container.

Overall, this is a stopgap measure to allow ttc to find helper
utilities needed for board manipulation.  However, since the docker
environment may be significantly different from the host, it may still
be missing something that is required for the helper utilities to run.

Signed-off-by: Tim Bird <tim.bird@sony.com>