docs: fix label for Adding_a_toolchain
Fix leftover tbwiki markup syntax for link to Adding_a_toolchain
in 'Adding_a_board.rst'.  Add an underscore to the label
to make it easier to read.

Also add a (commented out) option to conf.py to allow duplicate
labels between sections.  I'm not sure that this works and doesn't
cause other problems, so don't turn it on just yet.  But put
it there to avoid having to research it again (or to provide
a search term for when I work on this again.)

Signed-off-by: Tim Bird <tim.bird@sony.com>