docs: fix autosection labeling, and one new ref bug resulting from it
Add autosectionlabel_prefix_document to conf.py, so that duplicate
section names are unique-ified throughout the full document set.
This fixes about 120 warnings in the build, but introduces about 45
new warnings.  Start to fix those warnings (manually) by adding
a manual label for the criteria.json file (label=criteria_json).

The goal here is to fix up the new warnings manually, and get
to under 200 warnings for the doc build.

Signed-off-by: Tim Bird <tim.bird@sony.com>