fuego.conf: add use_binary_packages config option
This option instructs the system to use a binary package for
a test (if it exist), instead of doing the build and deploy
phases for the test.

Default this to 'false' for now.  It probabaly wouldn't
hurt to default to true, since you have to have packages in
your cache in order to activate this feature.  However,
this plays it safe, and makes sure that a user consciously
opts-in to this feature, by setting the value themselves

Signed-off-by: Tim Bird <tim.bird@sony.com>