jenkins: set more plugin values to fixed versions
Add jackson2-api and snakeyaml-api to the list of plugins
to pre-install with fixed versions.  Without this pre-installation,
the version that is installed is arbitrary.  This results
in many plugin failures, including description-setter, which
is used heavily by Fuego to set the build desription (including
the links to test artifacts)

Fix this issue by specifying fixed versions for these plugins.

Signed-off-by: Tim Bird <tim.bird@sony.com>