install: add --network host to docker build command
When behind a NAT or inside a firewall, the docker container
may not have sufficient network access to install debian
packages during a build.  Make the container build use the
networking on the host to overcome this problem.

This lack of network access was resulting in failure to resolve
hostname for debian servers, during the installation phase of
building the fuego container.

This fix seems to resolve the issue.

Signed-off-by: Tim Bird <tim.bird@sony.com>